Remove studentHolder Merge shell + config
