--- hw04-shell/src/main/java/ru/otus/hw/exceptions/AnonymousException.java
@@ -1,7 +1,0 @@
-package ru.otus.hw.exceptions;
-
-public class AnonymousException extends RuntimeException {
-    public AnonymousException(String message) {
-        super(message);
-    }
-}
