--- hw04-shell/src/main/java/ru/otus/hw/service/student/StudentHolder.java
@@ -1,13 +1,0 @@
-package ru.otus.hw.service.student;
-
-import ru.otus.hw.domain.Student;
-
-public interface StudentHolder {
-    Student getStudent();
-
-    void setStudent(Student student);
-
-    default boolean isStudentDetermined() {
-        return null != getStudent();
-    }
-}
