--- hw04-shell/src/main/java/ru/otus/hw/service/student/StudentHolderImpl.java
@@ -1,13 +1,0 @@
-package ru.otus.hw.service.student;
-
-import lombok.Data;
-import org.springframework.stereotype.Component;
-import ru.otus.hw.domain.Student;
-
-@Component
-@Data
-public class StudentHolderImpl implements StudentHolder {
-
-    private Student student;
-
-}
