+++ hw04-shell/src/main/java/ru/otus/hw/service/test/TestService.java
@@ -1,7 +1,8 @@
 package ru.otus.hw.service.test;
 
+import ru.otus.hw.domain.Student;
 import ru.otus.hw.domain.TestResult;
 
 public interface TestService {
-    TestResult executeTest();
+    TestResult executeTestFor(Student student);
 }
