+++ hw04-shell/src/test/java/ru/otus/hw/service/test/TestServiceImplTest.java
@@ -5,10 +5,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import ru.otus.hw.domain.TestResult;
-import ru.otus.hw.service.student.StudentHolder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static ru.otus.hw.service.test.TestConstants.QUESTION_RESULTS;
+import static ru.otus.hw.service.test.TestConstants.STUDENT;
 
 @DisplayName("Реализация TestService")
 @SpringBootTest(classes = TestConfig.class)
@@ -16,16 +16,14 @@ class TestServiceImplTest {
 
     @Autowired
     private TestService service;
-    @Autowired
-    private StudentHolder studentHolder;
 
     @Test
     @DisplayName("правильно строит результат")
     void rightResult() {
-        final var testResult = service.executeTest();
+        final var testResult = service.executeTestFor(STUDENT);
 
         assertThat(testResult)
-                .isEqualTo(new TestResult(studentHolder.getStudent(), QUESTION_RESULTS));
+                .isEqualTo(new TestResult(STUDENT, QUESTION_RESULTS));
     }
 
 }
